+++ service/key/error.go
@@ -8,3 +8,17 @@ var wrongTypeError = microerror.New("wrong type")
 func IsWrongTypeError(err error) bool {
 	return microerror.Cause(err) == wrongTypeError
 }
+
+var malformedCloudConfigKeyError = microerror.New("malformed key in the cloudconfig")
+
+// IsMalformedCloudConfigKey asserts malformedCloudConfigKeyError.
+func IsMalformedCloudConfigKey(err error) bool {
+	return microerror.Cause(err) == malformedCloudConfigKeyError
+}
+
+var missingCloudConfigKeyError = microerror.New("missing key in the cloudconfig")
+
+// IsMissingCloudConfigKey asserts missingCloudConfigKeyError.
+func IsMissingCloudConfigKey(err error) bool {
+	return microerror.Cause(err) == missingCloudConfigKeyError
+}
