+++ service/resource/legacy/load_balancer.go
@@ -10,7 +10,7 @@ import (
 
 	awsutil "github.com/giantswarm/aws-operator/client/aws"
 	awsresources "github.com/giantswarm/aws-operator/resources/aws"
-	"github.com/giantswarm/aws-operator/service/resource/cloudformation"
+	"github.com/giantswarm/aws-operator/service/key"
 )
 
 type LoadBalancerInput struct {
@@ -35,7 +35,7 @@ type LoadBalancerInput struct {
 }
 
 func (s *Resource) createLoadBalancer(input LoadBalancerInput) (*awsresources.ELB, error) {
-	lbName, err := cloudformation.LoadBalancerName(input.Name, input.Cluster)
+	lbName, err := key.LoadBalancerName(input.Name, input.Cluster)
 	if err != nil {
 		return nil, microerror.Mask(err)
 	}
@@ -79,7 +79,7 @@ func (s *Resource) createLoadBalancer(input LoadBalancerInput) (*awsresources.EL
 
 func (s *Resource) deleteLoadBalancer(input LoadBalancerInput) error {
 	// Delete ELB.
-	lbName, err := cloudformation.LoadBalancerName(input.Name, input.Cluster)
+	lbName, err := key.LoadBalancerName(input.Name, input.Cluster)
 	if err != nil {
 		return microerror.Mask(err)
 	}
