+++ service/resource/legacy/resource.go
@@ -24,7 +24,6 @@ import (
 	awsresources "github.com/giantswarm/aws-operator/resources/aws"
 	"github.com/giantswarm/aws-operator/service/cloudconfig"
 	"github.com/giantswarm/aws-operator/service/key"
-	"github.com/giantswarm/aws-operator/service/resource/cloudformation"
 )
 
 const (
@@ -1040,9 +1039,9 @@ func (s *Resource) processDelete(cluster awstpr.CustomObject) error {
 	}
 
 	// Delete Record Sets.
-	apiLBName, err := cloudformation.LoadBalancerName(cluster.Spec.Cluster.Kubernetes.API.Domain, cluster)
-	etcdLBName, err := cloudformation.LoadBalancerName(cluster.Spec.Cluster.Etcd.Domain, cluster)
-	ingressLBName, err := cloudformation.LoadBalancerName(cluster.Spec.Cluster.Kubernetes.IngressController.Domain, cluster)
+	apiLBName, err := key.LoadBalancerName(cluster.Spec.Cluster.Kubernetes.API.Domain, cluster)
+	etcdLBName, err := key.LoadBalancerName(cluster.Spec.Cluster.Etcd.Domain, cluster)
+	ingressLBName, err := key.LoadBalancerName(cluster.Spec.Cluster.Kubernetes.IngressController.Domain, cluster)
 	if err != nil {
 		s.logger.Log("error", fmt.Sprintf("%#v", err))
 	} else {
